+++ boot/uefi/efi_main.c
@@ -304,6 +304,8 @@ efi_status efi_main(efi_handle image_handle, efi_system_table *system_table) {
     info->magic = PEAK_BOOT_MAGIC;
     info->version = PEAK_BOOT_VERSION;
     info->flags = PEAK_BOOT_FLAG_HAS_FB;
+    if (conf.smoke_persist)
+        info->flags |= PEAK_BOOT_FLAG_SMOKE_PERSIST;
     info->dtb_phys = 0;
     info->dtb_size = 0;
     info->hhdm_offset = PEAK_HHDM_OFFSET;